requirements.txt
numpy
Cython
matplotlib
pandas
mysqlclient
optuna
scipy
scikit-posthocs
ECAgent
Pillow
argparse
animatplot